perf(udp-auth): pre-key HMAC, widen replay window
Pre-computed HMAC key schedule cloned per frame instead of re-keying:
skips 2 of the 5 SHA-256 compressions per 82-byte frame (measured seal
-46%, verify -57%).

Replay window 64 -> 1024 as a ring bitmap (WireGuard construction) so
advancing does not shift a map. 64 covered only 0.3s of the live 221
frame/s feed at max_age_ms=500, so a duplicate still fresh enough to
pass the staleness gate had already fallen out of the window.

Replay state is now keyed (key_id, peer) with a bounded 64-entry LRU:
two processes on one key_id (overlapping rolling forwarder deploy) own
independent sequence spaces, and sharing one window failed every frame
of the draining pod as outside-window.

Sender seed 0 instead of a ns wall clock, with verifier-side idle
expiry: a backward NTP step across a forwarder restart used to land
below the verifier's `highest` and reject permanently, clearable only
by restarting the RECEIVER. Restart recovery is now the verifier's job.

AuthReject enum replaces formatted strings so ingress can label
nxr_udp_auth_rejects_total{reason} without parsing prose. Drops the
dead `sequence != u64::MAX` guard (it tested the value already handed
out, so it could never prevent wrap).

+++ rust/src/pipeline_config.rs
@@ -555,8 +555,10 @@ pub struct NetworkYml {
 pub struct UdpAuthYml {
     pub max_age_ms: u64,
     pub max_future_ms: u64,
-    /// Sliding replay window (1..=64 sequence numbers).
-    pub replay_window: u8,
+    /// Sliding replay window (1..=1024 sequence numbers). MUST exceed
+    /// `max_age_ms` x peak frame rate, else a duplicate that is still fresh
+    /// enough to pass the staleness gate has already fallen out of the window.
+    pub replay_window: u16,
     pub keys: Vec<UdpAuthKeyYml>,
     /// Unauthenticated-frame disposition. Defaults to `strict` so an operator
     /// can never silently weaken ingest by omitting the field.
